requirements_rag.txt
# Core dependencies for the RAG system
openai>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Text processing and embeddings
tiktoken>=0.5.1
numpy>=1.24.0

# Vector storage options
faiss-cpu>=1.7.4  # For FAISS vector store (use faiss-gpu for GPU support)

# Database support
sqlite3  # Built into Python, but listed for clarity

# Optional: Enhanced text processing
# sentence-transformers>=2.2.2  # For alternative embedding models
# transformers>=4.35.0  # For local transformer models

# Optional: Advanced retrieval
# rank-bm25>=0.2.2  # For BM25 keyword search
# scikit-learn>=1.3.0  # For additional similarity metrics

# Development and testing
pytest>=7.4.0
httpx>=0.25.0  # For testing FastAPI endpoints 